+++ CHANGELOG.md
@@ -5,3 +5,32 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
 ## [Unreleased]
+
+## [1.0.8] - 2021-09-17
+### Fixed
+- Additional fixes for image upload, cache expiration and readme improvements [#10](https://github.com/rokwire/content-building-block/issues/11)
+
+## [1.0.7.3] - 2021-09-16
+### Fixed
+- Additional Twitter api improvements and fixes [#11](https://github.com/rokwire/content-building-block/issues/11)
+
+## [1.0.7.2] - 2021-09-15
+### Changed
+- Fixed missing cwebp executable within the docker image
+
+## [1.0.7.1] - 2021-09-15
+### Changed 
+- Fixed API doc & admin api security check[#10](https://github.com/rokwire/content-building-block/issues/10)
+
+## [1.0.7] - 2021-09-15
+### Added
+- Implement upload image api [#10](https://github.com/rokwire/content-building-block/issues/10)
+- Implement Twitter feed api [#11](https://github.com/rokwire/content-building-block/issues/11)
+## [1.0.6] - 2021-06-23
+### Fixed
+- GET /student_guides results "null" response on missing data in the table [#8](https://github.com/rokwire/content-building-block/issues/8)
+- Improve documentation [#6](https://github.com/rokwire/content-building-block/issues/6)
+
+### Added
+- Implement student guide client & admin APIs [#1](https://github.com/rokwire/content-building-block/issues/1)
+